% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{faceted.line.plot.dt}
\alias{faceted.line.plot.dt}
\title{Plot a stepped wedge design.}
\usage{
faceted.line.plot.dt(
  dataDT,
  studyDT,
  studyParams = NULL,
  ylims = NULL,
  outcomeName = "outcome.05",
  dotSize = 2,
  scatterColVar = c("site", "day")[[1]]
)
}
\arguments{
\item{dataDT}{Data from the SW study}

\item{studyDT}{Data about the SW study}

\item{ylims}{Limits of y-axis}

\item{outcomeName}{What outcome to plot}

\item{dotSize}{Size of points in scatter}

\item{scatterColVar}{What variable to colour the points on.}
}
\value{
The plot
}
\description{
Has facets for sites and clusters.
}
