% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.exp.R
\name{generate.trial}
\alias{generate.trial}
\title{Create a stepped wedge experiment structure.}
\usage{
generate.trial(nClusters, sitesPerCluster, timePerStep,
  transitionDuration = 0)
}
\arguments{
\item{nClusters}{How many clusters?}

\item{sitesPerCluster}{SHow many sites per cluster?}

\item{timePerStep}{How much time per step?}

\item{transitionDuration}{Study structure represented using classes from this package}
}
\value{
Stepped wedge experiment
}
\description{
Create a stepped wedge experiment structure.
}
