% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I.Sampler.R
\docType{class}
\name{I.Sampler-class}
\alias{I.Sampler-class}
\alias{I.Sampler}
\title{An interface for getting a value for a patient in a simulated context. Must
implement the get.value function.}
\description{
An interface for getting a value for a patient in a simulated context. Must
implement the get.value function.
}

