% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SW.Site.R
\docType{class}
\name{SW.Site-class}
\alias{SW.Site-class}
\alias{SW.Site}
\title{A Reference Class to represent a site in a stepped-wedge cluster-randomised
trial.}
\description{
A Reference Class to represent a site in a stepped-wedge cluster-randomised
trial.
}
\section{Fields}{

\describe{
\item{\code{name}}{The name of the site.}

\item{\code{sim.ppt.per.unit.time}}{For simulation purposes, the number of
participants generated per unit time.}

\item{\code{sim.normal.preintervention.mean}}{For simulation purposes, the mean
value of participants at the site before intervention.}

\item{\code{sim.normal.postintervention.mean}}{For simulation purposes, the mean
value of participants at the site after intervention.}

\item{\code{sim.normal.sd}}{For simulation purposes, the standard deviation of
participants at the site.}

\item{\code{sim.normal.effect.per.unit.time}}{For simulation purposes, the effect of
time on outcomes.}

\item{\code{timed.sampling.dist}}{A Timed.Sampling.Dist that can be used to simulate
non-normal data.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(name = "Site", randomisation.group = NA_integer_)}}{Initialize site with a name.}
}}

