% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SW.Study.R
\docType{class}
\name{SW.Study-class}
\alias{SW.Study-class}
\alias{SW.Study}
\title{A Reference Class to represent a stepped-wedge cluster-randomised trial.}
\description{
A Reference Class to represent a stepped-wedge cluster-randomised trial.
}
\section{Fields}{

\describe{
\item{\code{clusters}}{List of all clusters stored as a factor.}

\item{\code{study.dt}}{data.table that includes all site and cluster parameters
(some duplicated data).}

\item{\code{cluster.dt}}{data.table that includes cluster parameters \.}

\item{\code{site.dt}}{data.table that includes site parameters.}

\item{\code{data.dt}}{data.table that includes the data from participants.}

\item{\code{sim.ppt.per.unit.time.mean}}{Mean patients per unit time (will be forced
positive)}

\item{\code{sim.ppt.per.unit.time.sd}}{Variance in patients per unit time.}

\item{\code{sim.normal.preintervention.mean}}{The outcome before applying any
effects.}

\item{\code{sim.normal.intervention.effect.mean}}{The mean effect of the
intervention across all sites.}

\item{\code{sim.normal.intervention.effect.sd}}{The variance of the intervention
effect between sites.}

\item{\code{sim.site.effect.mean}}{The mean of the effect added separately to each
site.}

\item{\code{sim.site.effect.sd}}{The variance of the effect added separately to each
site.}

\item{\code{sim.site.effect.force.sign}}{Force the time effect to a certain sign
(should be in c(-1,0,+1))}

\item{\code{sim.time.effect.per.unit.mean}}{The mean of the effect added per unit
time for each site.}

\item{\code{sim.time.effect.per.unit.sd}}{The variance of the effect added per unit
time for each site.}

\item{\code{sim.individual.noise.mean}}{The mean of the variance between individuals
for each site.}

\item{\code{sim.individual.noise.sd}}{The variance of the variance between
individuals for each site.}

\item{\code{perm.dt}}{A table that holds a number of ways in which to permute sites
to different clusters.}

\item{\code{stat.dt}}{A statistic distribution generated by permuting sites to
different clusters.}
}}

\section{Methods}{

\describe{
\item{\code{add.clusters(clusters.to.add)}}{Add clusters to the experiment.}

\item{\code{cut.data.dt.time.to.periods(period.starts)}}{Cuts the continuous time variable in data.dt into periods, takes a vector of start times for cut}

\item{\code{faceted.line.plot(dot.size = 2, ylims = NULL, outcome.name = "outcome")}}{Plots the study as a line plot with facets for different sites. ylims and outcome.name are mainly for animations.}

\item{\code{generate.cluster.dt()}}{Generate a data.table of the clusters in the study.}

\item{\code{generate.perm.dt(max.r)}}{Wrapper for function generate.perm.dt in analysis.r}

\item{\code{generate.sim.data.normal(save.intermediates = T)}}{Generates synthetic data, can save the intermediate values for a nice animation if you want.}

\item{\code{generate.sim.data.tsd()}}{Generate data.dt from Timed.Sampling.Dist object.}

\item{\code{generate.site.dt()}}{Generate a data.table of the sites in the study.}

\item{\code{generate.site.sim.parameters()}}{Generate simultation parameters for each site.}

\item{\code{generate.stat.dt(
  max.r,
  outcome.col.name = "outcome",
  intervention.col.name = "group",
  stat.per.site = F,
  statistic = "WMWU",
  other.predictors = NULL,
  ...
)}}{Wrapper for function generate.stat.dt in analysis.r}

\item{\code{generate.study.dt()}}{Generate a data.table of the study.}

\item{\code{get.cluster.by.name(cluster.name)}}{Get a cluster given it's name.}

\item{\code{get.cluster.dt(force.generate = F)}}{Get a data.table of the clusters in the study.}

\item{\code{get.cluster.names()}}{Get all cluster names.}

\item{\code{get.clusters()}}{Get clusters in the study.}

\item{\code{get.site.by.name(site.name)}}{Get a site given it's name, needs to do a bit of trawling.}

\item{\code{get.site.dt(force.generate = F)}}{Get a data.table of the sites in the study.}

\item{\code{get.sites(cluster.name = "")}}{Gets a list of site objects, returning all sites for all cluster if no cluster.name provided.}

\item{\code{get.study.dt(force.generate = F)}}{Get a data.table of the study.}

\item{\code{rank.data.dt()}}{Gives ranks to outcomes in data.dt}

\item{\code{set.clusters(clusters)}}{Replace clusters in the cluster, clusters should be provided as a SW.Cluster object.}

\item{\code{set.sim.parameters(
  sim.ppt.per.unit.time.mean = NA_integer_,
  sim.ppt.per.unit.time.sd = NA_integer_,
  sim.normal.preintervention.mean = NA_integer_,
  sim.normal.intervention.effect.mean = NA_integer_,
  sim.normal.intervention.effect.sd = NA_integer_,
  sim.site.effect.mean = NA_integer_,
  sim.site.effect.sd = NA_integer_,
  sim.time.effect.per.unit.mean = NA_integer_,
  sim.time.effect.per.unit.sd = NA_integer_,
  sim.site.effect.force.sign = NA_integer_,
  sim.individual.noise.mean = NA_integer_,
  sim.individual.noise.sd = NA_integer_
)}}{Set simulation parameters for the study.}

\item{\code{set.sim.parameters.default(
  sim.ppt.per.unit.time.mean = 10,
  sim.ppt.per.unit.time.sd = 2,
  sim.normal.preintervention.mean = 15,
  sim.normal.intervention.effect.mean = 2,
  sim.normal.intervention.effect.sd = 0.6,
  sim.site.effect.mean = 0,
  sim.site.effect.sd = 2,
  sim.time.effect.per.unit.mean = 0,
  sim.time.effect.per.unit.sd = 0.005,
  sim.site.effect.force.sign = c(-1, 0, +1)[3],
  sim.individual.noise.mean = 2,
  sim.individual.noise.sd = 0.4
)}}{Set simulation parameters for the study.}

\item{\code{update.groups.data.dt()}}{Updates the group of data.dt according to timing of clusters.}
}}

\section{Normal simulation}{
NA
}

\section{Permutation testing}{
NA
}

