% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{test.stat.table.per.site}
\alias{test.stat.table.per.site}
\title{Applies a few tests to a statistic table, generated by generate.stat.dt I
don't think this is as good as an overall wilcoxon}
\usage{
test.stat.table.per.site(stat.dt)
}
\arguments{
\item{stat.dt}{A data.table of the statistic with minimally columns for stat,
permuted, and I guess site if you want that in the plot}
}
\value{
A list with a few tests.
}
\description{
Applies a few tests to a statistic table, generated by generate.stat.dt I
don't think this is as good as an overall wilcoxon
}
