% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{test.stat.table}
\alias{test.stat.table}
\title{Applies a few tests to a statistic table, generated by generate.stat.dt}
\usage{
test.stat.table(stat.dt)
}
\arguments{
\item{stat.dt}{A data.table of the statistic with minimally columns for stat,
permuted.}
}
\value{
A list with only a p-value currently.
}
\description{
Applies a few tests to a statistic table, generated by generate.stat.dt
}
