% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{test.wilcoxon.ANOVA.form}
\alias{test.wilcoxon.ANOVA.form}
\title{Applies a two-way ANOVA to Wilcoxon ranks}
\usage{
test.wilcoxon.ANOVA.form(data.dt, form)
}
\arguments{
\item{data.dt}{A data.table with results from each participant, an outcome, a
group (i.e. effect of interest), and a time}
}
\value{
Results of the two-way ANOVA.
}
\description{
Applies a two-way ANOVA to Wilcoxon ranks
}
