% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{test.wilcoxon.ANOVA}
\alias{test.wilcoxon.ANOVA}
\title{Applies a two-way ANOVA to Wilcoxon ranks}
\usage{
test.wilcoxon.ANOVA(
  data.dt,
  outcome.col.name = "outcome",
  intervention.col.name = "group",
  other.predictors = "time"
)
}
\arguments{
\item{data.dt}{A data.table with results from each participant, an outcome, a
group (i.e. effect of interest), and a time}
}
\value{
Results of the two-way ANOVA.
}
\description{
Applies a two-way ANOVA to Wilcoxon ranks
}
