% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timed.Sampling.Dist.R
\docType{class}
\name{Timed.Sampling.Dist-class}
\alias{Timed.Sampling.Dist-class}
\alias{Timed.Sampling.Dist}
\title{A Reference Class to represent a sampling distribution that depends on a time
that is provided.}
\description{
A Reference Class to represent a sampling distribution that depends on a time
that is provided.
}
\section{Fields}{

\describe{
\item{\code{sampling.dt}}{A data.table with three columns time.begin/time.end (time when the
sampling distribution starts(>=)/stops(<), NA means ignore), weight
(probabilities), and value (the value to be chosen).}

\item{\code{legal.col.names}}{A vector of names that are allowed as column names for
sampling.dt}
}}

\section{Methods}{

\describe{
\item{\code{add.sampling.dist(
  values,
  weights = c(),
  time.begin = NA_real_,
  time.end = NA_real_
)}}{Adds a sampling distribution at the specified time. If either time limit is NA, there is no limit. Weights and values should be vectors of same length.}

\item{\code{get.quantile.dt(at.quantile = 0.5)}}{Gives the quantiles at all time periods, at at.quantile (e.g. 0.5 for median).}

\item{\code{merge.periods.between(merge.time.begin, merge.time.end)}}{Merges all periods between time.begin (default begin first period) and time.end (default end last period). Weights will be average of composing periods.}

\item{\code{normalise.sampling.dt(weights.sum = 1000)}}{Makes the weights add to a certain value.}

\item{\code{set.sampling.dt(input.sampling.dt)}}{Sets entire sampling data.table, remember to give it the right columns!}

\item{\code{sort.sampling.dt()}}{Sorts sampling dt by time}

\item{\code{split.period(time.split, silent = F)}}{Split the distribution at a certain time, resulting in two identical distributions either side, the second starting at time.split (unless you choose to split at an existing split, in which case nothing happens}
}}

