% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{generate.perm.dt}
\alias{generate.perm.dt}
\title{Generate a table that holds a number of ways in which to permute sites to
different clusters.}
\usage{
generate.perm.dt(study.dt, max.r = 1000)
}
\arguments{
\item{study.dt}{A study info data.table, containing site, cluster, and timing info.}

\item{max.r}{Number of permutations.}
}
\value{
A data.table containing all the permutations of site to cluster.
}
\description{
Generate a table that holds a number of ways in which to permute sites to
different clusters.
}
