% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pseudo_median}
\alias{pseudo_median}
\title{Cut down version of wilcox.test that only returns the pseudo-median without
CIs. Uses Rfast for a little bit of a speed boost.}
\usage{
pseudo_median(x)
}
\arguments{
\item{x}{Vector of values.}
}
\value{
Pseudo-median of x
}
\description{
Cut down version of wilcox.test that only returns the pseudo-median without
CIs. Uses Rfast for a little bit of a speed boost.
}
